Break up setup into forcecalculator and integrator
