+++ src/gromacs/nblib/forcecalculator.cpp
@@ -40,12 +40,104 @@
  */
 
 #include "forcecalculator.h"
+#include "integrator.h"
+
+#include "gromacs/compat/optional.h"
+#include "gromacs/ewald/ewald_utils.h"
+#include "gromacs/gmxlib/nrnb.h"
+#include "gromacs/mdlib/dispersioncorrection.h"
+#include "gromacs/mdlib/force_flags.h"
+#include "gromacs/mdlib/forcerec.h"
+#include "gromacs/mdlib/gmx_omp_nthreads.h"
+#include "gromacs/mdtypes/enerdata.h"
+#include "gromacs/mdtypes/forcerec.h"
+#include "gromacs/mdtypes/interaction_const.h"
+#include "gromacs/mdtypes/mdatom.h"
+#include "gromacs/mdtypes/simulation_workload.h"
+#include "gromacs/nbnxm/atomdata.h"
+#include "gromacs/nbnxm/gridset.h"
+#include "gromacs/nbnxm/nbnxm.h"
+#include "gromacs/nbnxm/nbnxm_simd.h"
+#include "gromacs/nbnxm/pairlistset.h"
+#include "gromacs/nbnxm/pairlistsets.h"
+#include "gromacs/nbnxm/pairsearch.h"
+#include "gromacs/pbcutil/ishift.h"
+#include "gromacs/pbcutil/mshift.h"
+#include "gromacs/pbcutil/pbc.h"
+#include "gromacs/simd/simd.h"
+#include "gromacs/timing/cyclecounter.h"
+#include "gromacs/topology/topology.h"
+#include "gromacs/utility/enumerationhelpers.h"
+#include "gromacs/utility/fatalerror.h"
+#include "gromacs/utility/logger.h"
+
+
+//! Print timings outputs
+void printTimingsOutput(const NBKernelOptions &options,
+                               const NBKernelSystem  &system,
+                               const gmx::index      &numPairs,
+                               gmx_cycles_t           cycles)
+{
+    const gmx::EnumerationArray<BenchMarkKernels, std::string>  kernelNames   = { "auto", "no", "4xM", "2xMM" };
+    const gmx::EnumerationArray<BenchMarkCombRule, std::string> combruleNames = { "geom.", "LB", "none" };
+
+    // Generate an, accurate, estimate of the number of non-zero pair interactions
+    const real                          atomDensity          = system.coordinates.size()/det(system.box);
+    const real                          numPairsWithinCutoff = atomDensity*4.0/3.0*M_PI*std::pow(options.pairlistCutoff, 3);
+    const real                          numUsefulPairs       = system.coordinates.size()*0.5*(numPairsWithinCutoff + 1);
+#if GMX_SIMD
+    if (options.nbnxmSimd != BenchMarkKernels::SimdNo)
+    {
+        fprintf(stdout, "SIMD width:           %d\n", GMX_SIMD_REAL_WIDTH);
+    }
+#endif
+    fprintf(stdout, "System size:          %zu atoms\n", system.coordinates.size());
+    fprintf(stdout, "Cut-off radius:       %g nm\n", options.pairlistCutoff);
+    fprintf(stdout, "Number of threads:    %d\n", options.numThreads);
+    fprintf(stdout, "Number of iterations: %d\n", options.numIterations);
+    fprintf(stdout, "Compute energies:     %s\n",
+            options.computeVirialAndEnergy ? "yes" : "no");
+    if (options.coulombType != BenchMarkCoulomb::ReactionField)
+    {
+        fprintf(stdout, "Ewald excl. corr.:    %s\n",
+                options.nbnxmSimd == BenchMarkKernels::SimdNo || options.useTabulatedEwaldCorr ? "table" : "analytical");
+    }
+    printf("\n");
+
+    fprintf(stdout, "Coulomb LJ   comb. SIMD    Mcycles  Mcycles/it.   %s\n",
+            options.cyclesPerPair ? "cycles/pair" : "pairs/cycle");
+    fprintf(stdout, "                                                total    useful\n");
+
+    fprintf(stdout, "%-7s %-4s %-5s %-4s ",
+            options.coulombType == BenchMarkCoulomb::Pme ? "Ewald" : "RF",
+            options.useHalfLJOptimization ? "half" : "all",
+            combruleNames[options.ljCombinationRule].c_str(),
+            kernelNames[options.nbnxmSimd].c_str());
+
+    const double dCycles = static_cast<double>(cycles);
+    if (options.cyclesPerPair)
+    {
+        fprintf(stdout, "%10.3f %10.4f %8.4f %8.4f\n",
+                cycles*1e-6,
+                dCycles/options.numIterations*1e-6,
+                dCycles/(options.numIterations*numPairs),
+                dCycles/(options.numIterations*numUsefulPairs));
+    }
+    else
+    {
+        fprintf(stdout, "%10.3f %10.4f %8.4f %8.4f\n",
+                dCycles*1e-6,
+                dCycles/options.numIterations*1e-6,
+                options.numIterations*numPairs/dCycles,
+                options.numIterations*numUsefulPairs/dCycles);
+    }
+}
 
 
 //! Sets up and runs the kernel calls
 //! TODO Refactor this function to return a handle to dispatchNonbondedKernel
 //!      that callers can manipulate directly.
-static void setupAndRunInstance(NBKernelSystem          &system,
+void setupAndRunInstance(NBKernelSystem          &system,
                                 const NBKernelOptions   &options,
                                 const bool              &printTimings)
 {
