--- src/gromacs/nblib/interactions.cpp
@@ -1,5 +1,0 @@
-//
-// Created by sebkelle on 20.11.19.
-//
-
-#include "interactions.h"
