+++ src/gromacs/nblib/setup.cpp
@@ -45,7 +45,8 @@
 #include "gmxpre.h"
 
 #include "setup.h"
-#include "integrator.h"
+// #include "integrator.h"
+// #include "forcecalculator.h"
 
 #include "gromacs/compat/optional.h"
 #include "gromacs/ewald/ewald_utils.h"
@@ -55,7 +56,6 @@
 #include "gromacs/mdlib/forcerec.h"
 #include "gromacs/mdlib/gmx_omp_nthreads.h"
 #include "gromacs/mdtypes/enerdata.h"
-#include "gromacs/mdtypes/forcerec.h"
 #include "gromacs/mdtypes/interaction_const.h"
 #include "gromacs/mdtypes/mdatom.h"
 #include "gromacs/mdtypes/simulation_workload.h"
@@ -139,8 +139,7 @@ static real ewaldCoeff(const real ewald_rtol, const real pairlistCutoff)
 }
 
 //! Return an interaction constants struct with members used in the benchmark set appropriately
-static interaction_const_t setupInteractionConst(const NBKernelOptions &options)
-
+interaction_const_t setupInteractionConst(const NBKernelOptions &options)
 {
     interaction_const_t ic;
 
@@ -170,7 +169,7 @@ static interaction_const_t setupInteractionConst(const NBKernelOptions &options)
 }
 
 //! Sets up and returns a Nbnxm object for the given options and system
-static std::unique_ptr<nonbonded_verlet_t>
+std::unique_ptr<nonbonded_verlet_t>
 setupNbnxmInstance(const NBKernelOptions   &options,
                    NBKernelSystem          &system)
 {
@@ -278,66 +277,6 @@ expandSimdOptionAndPushBack(const NBKernelOptions        &options,
     }
 }
 
-//! Print timings outputs
-static void printTimingsOutput(const NBKernelOptions &options,
-                               const NBKernelSystem  &system,
-                               const gmx::index      &numPairs,
-                               gmx_cycles_t           cycles)
-{
-    const gmx::EnumerationArray<BenchMarkKernels, std::string>  kernelNames   = { "auto", "no", "4xM", "2xMM" };
-    const gmx::EnumerationArray<BenchMarkCombRule, std::string> combruleNames = { "geom.", "LB", "none" };
-
-    // Generate an, accurate, estimate of the number of non-zero pair interactions
-    const real                          atomDensity          = system.coordinates.size()/det(system.box);
-    const real                          numPairsWithinCutoff = atomDensity*4.0/3.0*M_PI*std::pow(options.pairlistCutoff, 3);
-    const real                          numUsefulPairs       = system.coordinates.size()*0.5*(numPairsWithinCutoff + 1);
-#if GMX_SIMD
-    if (options.nbnxmSimd != BenchMarkKernels::SimdNo)
-    {
-        fprintf(stdout, "SIMD width:           %d\n", GMX_SIMD_REAL_WIDTH);
-    }
-#endif
-    fprintf(stdout, "System size:          %zu atoms\n", system.coordinates.size());
-    fprintf(stdout, "Cut-off radius:       %g nm\n", options.pairlistCutoff);
-    fprintf(stdout, "Number of threads:    %d\n", options.numThreads);
-    fprintf(stdout, "Number of iterations: %d\n", options.numIterations);
-    fprintf(stdout, "Compute energies:     %s\n",
-            options.computeVirialAndEnergy ? "yes" : "no");
-    if (options.coulombType != BenchMarkCoulomb::ReactionField)
-    {
-        fprintf(stdout, "Ewald excl. corr.:    %s\n",
-                options.nbnxmSimd == BenchMarkKernels::SimdNo || options.useTabulatedEwaldCorr ? "table" : "analytical");
-    }
-    printf("\n");
-
-    fprintf(stdout, "Coulomb LJ   comb. SIMD    Mcycles  Mcycles/it.   %s\n",
-            options.cyclesPerPair ? "cycles/pair" : "pairs/cycle");
-    fprintf(stdout, "                                                total    useful\n");
-
-    fprintf(stdout, "%-7s %-4s %-5s %-4s ",
-            options.coulombType == BenchMarkCoulomb::Pme ? "Ewald" : "RF",
-            options.useHalfLJOptimization ? "half" : "all",
-            combruleNames[options.ljCombinationRule].c_str(),
-            kernelNames[options.nbnxmSimd].c_str());
-
-    const double dCycles = static_cast<double>(cycles);
-    if (options.cyclesPerPair)
-    {
-        fprintf(stdout, "%10.3f %10.4f %8.4f %8.4f\n",
-                cycles*1e-6,
-                dCycles/options.numIterations*1e-6,
-                dCycles/(options.numIterations*numPairs),
-                dCycles/(options.numIterations*numUsefulPairs));
-    }
-    else
-    {
-        fprintf(stdout, "%10.3f %10.4f %8.4f %8.4f\n",
-                dCycles*1e-6,
-                dCycles/options.numIterations*1e-6,
-                options.numIterations*numPairs/dCycles,
-                options.numIterations*numUsefulPairs/dCycles);
-    }
-}
 
 
 void nbKernel(NBKernelSystem        &system,
@@ -362,5 +301,5 @@ void nbKernel(NBKernelSystem        &system,
     expandSimdOptionAndPushBack(options, &optionsList);
     GMX_RELEASE_ASSERT(!optionsList.empty(), "Expect at least one benchmark setup");
 
-    setupAndRunInstance(system, optionsList[0], printTimings);
+    // setupAndRunInstance(system, optionsList[0], printTimings);
 }
