+++ src/gromacs/nblib/tests/CMakeLists.txt
@@ -34,7 +34,7 @@
 
 gmx_add_unit_test(GmxNBLibTests nblib
     box.cpp
-    nblib.cpp
+    #nblib.cpp
     #simstate.cpp
     util.cpp
     )
